⌛Refactor: of, from 메서드 위치수정
- Entity의 of, from 메서드 DTO로 옮김
- Service 내부 코드도 이에 맞게 수정
